+++ src/components/Dropdown/DropdownBlockLink/index.js
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import Checkbox from '../../Checkbox';
 
-export const DropdownBlockLink = (props) => {
+export const DropdownBlockLink = props => {
   const el = useRef(null);
 
   useEffect(() => {
@@ -19,6 +19,14 @@ export const DropdownBlockLink = (props) => {
     props.onClick(props.value);
   }
 
+  function onMouseEnter() {
+    props.onMouseEnter(props.value);
+  }
+
+  function onMouseLeave() {
+    props.onMouseLeave(props.value);
+  }
+
   const styleProps = {};
   if (props.minWidth) {
     styleProps.minWidth = props.minWidth;
@@ -27,8 +35,8 @@ export const DropdownBlockLink = (props) => {
   return (
     <div
       className={ classNames({
-        'link': props.isLink,
-        'isSelected': !props.isLink,
+        link: props.isLink,
+        isSelected: !props.isLink,
         'oui-dropdown__block-link': props.isLink,
         'oui-dropdown__block-link--has-focus': props.hasFauxFocus,
       }) }
@@ -36,16 +44,14 @@ export const DropdownBlockLink = (props) => {
       style={ styleProps }
       { ...(props.testSection ? { 'data-test-section': props.testSection } : {}) }
       { ...(props.trackId ? { 'data-track-id': props.trackId } : {}) }
-      onClick={ onClick }>
-      {
-        props.isMultiSelect ?
-          <Checkbox
-            defaultChecked={ props.isItemSelected }
-            label={ props.children }
-            isDisabled={ false }
-          /> :
-          props.children
-      }
+      onClick={ onClick }
+      onMouseEnter={ onMouseEnter }
+      onMouseLeave={ onMouseLeave }>
+      {props.isMultiSelect ? (
+        <Checkbox defaultChecked={ props.isItemSelected } label={ props.children } isDisabled={ false } />
+      ) : (
+        props.children
+      )}
     </div>
   );
 };
@@ -70,13 +76,14 @@ DropdownBlockLink.propTypes = {
   /** Whether or not this item should include a checkbox */
   isMultiSelect: PropTypes.bool,
   /** Minimum width of the list item, useful if
-    * you need to have a block of description text */
-  minWidth: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]),
+   * you need to have a block of description text */
+  minWidth: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   /** Click handler for the menu option */
   onClick: PropTypes.func.isRequired,
+  /** MouseEnter handler for the menu option */
+  onMouseEnter: PropTypes.func,
+  /** MouseLeave handler for the menu option */
+  onMouseLeave: PropTypes.func,
   /** Used for data-test-section attribute on the link */
   testSection: PropTypes.string,
   /** Used for data-track-id attribute on the link */